+++ rmk/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix with_modifier action doesn't trigger the key with modifier
+
 ## [0.2.2] - 2024-07-12
 
 - Add keyboard macro support