+++ rmk/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Update versions of dependecies
+
 ## [0.2.3] - 2024-07-25
 
 ### Fixed